MassCardView
Things changed:

* Added card stack
* Standardized damage to ints rather than floats.

At some point, we're going to want to make the mass card view more helpful by highlighting cards in different locations. Like if they're in your hand, they have a different colored border.

Note: I was planning on having a 'banished' pile, which is similar to the consume pile in Monster Train, or exile from MTG. Cards with "Banish:X" on them will go to the banished zone when played and stay there until you clear X floors.

+++ Assets/Scripts/Player/Cards/CardImplementations/CardCinders.cs
@@ -5,7 +5,7 @@ using UnityEngine.Tilemaps;
 
 public class CardCinders : Card
 {
-    float damage = 10f;
+    int damage = 10 ;
     public CardCinders()
     {
         cardName = "Cinders";
